luong_nghin_do/requirements.txt
django
djangorestframework
mysqlclient
openai
djangorestframework
mysqlclient
python-docx 
pymupdf 
python-pptx
speechrecognition 
pydub
gTTS
openai
cloudinary
django-cors-headers
coverage
